# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AquaFarm Pro API
  version: 0.2.0
  description: |
    Core service contract for AquaFarm Pro.

    Included domains:
      - Authentication (JWT issuance & profile)
      - Multi‑tenancy (tenant creation / scoping)
      - Farms & Ponds (infrastructure inventory)
      - Water Quality (readings + trend analysis)
      - Fish Batches (stocking, growth lifecycle)
      - Feeding Records (nutrition tracking)
      - Accounting (journal entry draft capability)

    Notable design conventions:
      - Pagination envelope: { data: T[], meta } (all list endpoints unified; legacy array root keys deprecated and scheduled for removal in v0.4.0)
      - Standardized error responses via reusable components (ValidationError, UnauthorizedError, NotFoundError, ForbiddenError)
      - Consistent 201 bodies & 204 for destructive deletes
      - Examples supplied for major resource schemas & list responses

    Change log highlights:
      - 0.2.0: Added reusable list response components, standardized error $refs, multi-scenario water quality examples, enriched metadata.
      - 0.1.0: Initial rebuilt contract after corruption (auth, tenants, farms, ponds, core water quality & fish batches).
  termsOfService: https://example.com/terms
  contact:
    name: AquaFarm API Support
    url: https://example.com/support
    email: api-support@example.com
  license:
    name: UNLICENSED
    url: https://example.com/license
servers:
  - url: https://api.aquafarm.local/v1
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Tenant:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        locale:
          type: string
...
    Pond:
      type: object
      description: Pond resource inside a farm.
      required: [id, name, farmId, tenantId, area, depth, volume, maxCapacity]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        area: { type: number, format: float, description: Surface area (m²) }
        depth: { type: number, format: float, description: Average depth (m) }
        volume: { type: number, format: float, description: Volume (m³) }
        maxCapacity: { type: integer, description: Maximum fish capacity }
        currentStockCount: { type: integer, description: Current fish count }
        shape: { type: string, enum: [rectangular, circular, irregular], default: rectangular }
        status: { type: string, enum: [active, maintenance, inactive, cleaning], default: active }
        equipment:
          type: array
          items: { type: string }
          description: Installed equipment identifiers
        coordinates:
          type: object
          nullable: true
          properties:
            latitude: { type: number, format: float }
            longitude: { type: number, format: float }
        notes: { type: string, nullable: true }
        farmId: { type: string, format: uuid }
        managedById: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        utilizationRate: { type: number, readOnly: true }
        stockingDensity: { type: number, readOnly: true }
        isOverstocked: { type: boolean, readOnly: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        tenant_id:
          type: string
          description: Deprecated alias for tenantId
          deprecated: true
      example:
        id: "e6d0d8a2-5c2f-4ca2-8b6b-1d9f6b6e9a11"
        name: Growout Pond 3
        description: Primary tilapia production pond
        area: 1200
        depth: 1.8
        volume: 2160
        maxCapacity: 8000
        currentStockCount: 6200
        shape: rectangular
        status: active
        equipment: ["aerator-01", "sensor-kit-07"]
        coordinates:
          latitude: 14.6234
          longitude: 121.0123
        notes: Weekly sampling on Mondays
        farmId: "5a5935aa-fd8a-433e-9b6d-111b8e610001"
        managedById: "c4f0b5c1-8ff1-43e2-9e09-4321a2f90001"
        tenantId: "92f7c9f0-1111-4c52-a5cd-9a11e3b90001"
        utilizationRate: 0.78
        stockingDensity: 5.17
        isOverstocked: false
        createdAt: "2025-09-01T10:15:30Z"
        updatedAt: "2025-09-15T08:42:11Z"
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string }
        tenantId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
...
    WaterQualityReading:
      type: object
      description: A single captured water quality reading for a pond.
      required:
        [id, pondId, temperature, ph, dissolvedOxygen, ammonia, nitrite, nitrate, status, recordedById, createdAt]
      properties:
        id: { type: string, format: uuid }
        pondId: { type: string, format: uuid }
        recordedById: { type: string, format: uuid }
        temperature: { type: number, format: float, description: Water temperature in °C }
        ph: { type: number, format: float }
        dissolvedOxygen: { type: number, format: float, description: mg/L }
        ammonia: { type: number, format: float, description: mg/L }
        nitrite: { type: number, format: float, description: mg/L }
        nitrate: { type: number, format: float, description: mg/L }
        salinity: { type: number, format: float, nullable: true, description: ppt }
        turbidity: { type: number, format: float, nullable: true, description: NTU }
        alkalinity: { type: number, format: float, nullable: true }
        hardness: { type: number, format: float, nullable: true }
        readingMethod:
          type: string
          enum: [manual, automatic, sensor]
          default: manual
        notes: { type: string, nullable: true }
        status:
          type: string
          enum: [normal, warning, critical]
          default: normal
        alerts:
          type: array
          items: { type: string }
          description: Machine-generated alert codes based on threshold analysis
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      example:
        id: "c0b7d5e9-1fb8-41d9-b5bb-a9b5a0c00001"
        pondId: "e6d0d8a2-5c2f-4ca2-8b6b-1d9f6b6e9a11"
        recordedById: "c4f0b5c1-8ff1-43e2-9e09-4321a2f90001"
        temperature: 27.4
        ph: 7.6
        dissolvedOxygen: 6.8
        ammonia: 0.02
        nitrite: 0.01
        nitrate: 10.5
        salinity: null
        turbidity: 4.2
        alkalinity: 120
        hardness: 150
        readingMethod: sensor
        notes: Morning reading
        status: normal
        alerts: []
        createdAt: "2025-09-15T06:45:11Z"
        updatedAt: "2025-09-15T06:45:11Z"
      x-examples:
        normal:
          summary: All parameters within target thresholds
          value:
            id: "c0b7d5e9-1fb8-41d9-b5bb-a9b5a0c00001"
            pondId: "e6d0d8a2-5c2f-4ca2-8b6b-1d9f6b6e9a11"
            recordedById: "c4f0b5c1-8ff1-43e2-9e09-4321a2f90001"
            temperature: 27.4
            ph: 7.6
            dissolvedOxygen: 6.8
            ammonia: 0.02
            nitrite: 0.01
            nitrate: 10.5
            status: normal
            alerts: []
            createdAt: "2025-09-15T06:45:11Z"
            updatedAt: "2025-09-15T06:45:11Z"
        warning:
            summary: Dissolved oxygen trending low
            value:
              id: "d1d2e3f4-2fb8-41d9-b5bb-a9b5a0c00002"
              pondId: "e6d0d8a2-5c2f-4ca2-8b6b-1d9f6b6e9a11"
              recordedById: "c4f0b5c1-8ff1-43e2-9e09-4321a2f90001"
              temperature: 29.1
              ph: 7.4
              dissolvedOxygen: 4.5
              ammonia: 0.03
              nitrite: 0.02
              nitrate: 14.2
              status: warning
              alerts: ["LOW_DO_TREND"]
              createdAt: "2025-09-15T12:15:22Z"
              updatedAt: "2025-09-15T12:15:22Z"
        critical:
            summary: Multiple critical thresholds exceeded
            value:
              id: "e2e3f4a5-2fb8-41d9-b5bb-a9b5a0c00003"
              pondId: "e6d0d8a2-5c2f-4ca2-8b6b-1d9f6b6e9a11"
              recordedById: "c4f0b5c1-8ff1-43e2-9e09-4321a2f90001"
              temperature: 31.0
              ph: 6.5
              dissolvedOxygen: 3.2
              ammonia: 0.20
              nitrite: 0.10
              nitrate: 40.0
              status: critical
              alerts: ["LOW_DO_CRITICAL", "AMMONIA_HIGH", "NITRITE_HIGH"]
              createdAt: "2025-09-15T14:30:05Z"
              updatedAt: "2025-09-15T14:30:05Z"
    CreateWaterQualityReadingRequest:
      type: object
      required: [pondId, temperature, ph, dissolvedOxygen, ammonia, nitrite, nitrate]
      properties:
        pondId: { type: string, format: uuid }
        temperature: { type: number }
        ph: { type: number }
        dissolvedOxygen: { type: number }
        ammonia: { type: number }
        nitrite: { type: number }
        nitrate: { type: number }
        salinity: { type: number, nullable: true }
        turbidity: { type: number, nullable: true }
        alkalinity: { type: number, nullable: true }
        hardness: { type: number, nullable: true }
        readingMethod: { type: string, enum: [manual, automatic, sensor] }
        notes: { type: string, nullable: true }
...
    PaginatedPondResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Pond' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }
...
    PaginatedUserResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/User' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }
...
  parameters:
    XTenantIdHeader:
      name: X-Tenant-Id
      in: header
      required: false
      description: Tenant code or UUID for scoping requests
      schema: { type: string }
    PageQuery:
      name: page
      in: query
      required: false
      description: |
        Page number (1-based). If the requested page exceeds the last page, the server clamps it to the last available page and returns that page's meta (no error raised).
      schema: { type: integer, minimum: 1, default: 1 }
    PondIdQuery:
      name: pondId
      in: query
      required: false
      description: Filter readings by pond ID
      schema: { type: string, format: uuid }
    LimitQuery:
      name: limit
      in: query
      required: false
      description: |
        Page size (items per page). Values > 100 are capped to 100. Values < 1 are normalized to 1. Default varies by endpoint (commonly 25).
      schema: { type: integer, minimum: 1, maximum: 100 }
    DaysQuery:
      name: days
      in: query
      required: false
      description: Number of past days to analyze (default 7)
      schema: { type: integer, minimum: 1, maximum: 90 }
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    PondListResponse:
      description: Paginated list of ponds
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PaginatedPondResponse'
          example:
            data:
              - id: "e6d0d8a2-5c2f-4ca2-8b6b-1d9f6b6e9a11"
                name: Growout Pond 3
                area: 1200
                depth: 1.8
                volume: 2160
                maxCapacity: 8000
                currentStockCount: 6200
                shape: rectangular
                status: active
                farmId: "5a5935aa-fd8a-433e-9b6d-111b8e610001"
                tenantId: "92f7c9f0-1111-4c52-a5cd-9a11e3b90001"
                createdAt: "2025-09-01T10:15:30Z"
                updatedAt: "2025-09-15T08:42:11Z"
            meta:
              total: 1
              page: 1
              limit: 10
              totalPages: 1
              hasNext: false
              hasPrev: false
...
    WaterQualityReadingListResponse:
      description: Paginated list of water quality readings
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PaginatedWaterQualityReadingResponse'
          example:
            data:
              - id: "c0b7d5e9-1fb8-41d9-b5bb-a9b5a0c00001"
                pondId: "e6d0d8a2-5c2f-4ca2-8b6b-1d9f6b6e9a11"
                temperature: 27.4
                ph: 7.6
                dissolvedOxygen: 6.8
                ammonia: 0.02
                nitrite: 0.01
                nitrate: 10.5
                status: normal
                alerts: []
                recordedById: "c4f0b5c1-8ff1-43e2-9e09-4321a2f90001"
                createdAt: "2025-09-15T06:45:11Z"
                updatedAt: "2025-09-15T06:45:11Z"
            meta:
              total: 1
              page: 1
              limit: 10
              totalPages: 1
              hasNext: false
              hasPrev: false
...
    UserListResponse:
      description: Paginated list of users
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PaginatedUserResponse'
          example:
            data:
              - id: "11111111-2222-3333-4444-555555555555"
                name: Admin User
                email: admin@aquafarm.local
                role: admin
                tenantId: "92f7c9f0-1111-4c52-a5cd-9a11e3b90001"
                createdAt: "2025-09-01T09:00:00Z"
            meta:
              total: 1
              page: 1
              limit: 10
              totalPages: 1
              hasNext: false
              hasPrev: false
...
paths:
...
  /tenants:
    get:
      tags: [Tenants]
      summary: List tenants (admin scope)
      operationId: listTenants
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tenant'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
    post:
      tags: [Tenants]
      summary: Create tenant
      operationId: createTenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                locale:
                  type: string
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
  /ponds:
    get:
      tags: [Ponds]
      summary: List ponds for current tenant
      operationId: listPonds
      parameters:
        - $ref: '#/components/parameters/XTenantIdHeader'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200': { $ref: '#/components/responses/PondListResponse' }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags: [Ponds]
      summary: Create pond
      operationId: createPond
      parameters:
        - $ref: '#/components/parameters/XTenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [farmId, name, area, depth, maxCapacity]
              properties:
                farmId: { type: string, format: uuid }
                name: { type: string }
                area: { type: number }
                depth: { type: number }
                maxCapacity: { type: integer }
                shape: { type: string }
      responses:
        '201':
          description: Pond created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pond'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /ponds/{id}:
    get:
      tags: [Ponds]
      summary: Get pond by id
      operationId: getPond
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/XTenantIdHeader'
      responses:
        '200':
          description: Pond details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pond'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    patch:
      tags: [Ponds]
      summary: Update pond
      operationId: updatePond
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/XTenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                area: { type: number }
                depth: { type: number }
                maxCapacity: { type: integer }
                status: { type: string }
      responses:
        '200': { description: Pond updated }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    delete:
      tags: [Ponds]
      summary: Delete pond
      operationId: deletePond
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/XTenantIdHeader'
      responses:
        '204': { description: Pond deleted }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
...
  /users:
    get:
      tags: [Auth]
      summary: List users (scoped to tenant unless admin)
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200': { $ref: '#/components/responses/UserListResponse' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
    post:
      tags: [Auth]
      summary: Create user (admin only in future; currently open for internal bootstrap/testing)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string }
                role: { type: string }
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400': { $ref: '#/components/responses/ValidationError' }
  /users/{id}:
    get:
      tags: [Auth]
      summary: Get user by id
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    patch:
      tags: [Auth]
      summary: Update user by id
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                role: { type: string }
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    delete:
      tags: [Auth]
      summary: Delete user by id
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: User deleted }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /water-quality:
    get:
      tags: [WaterQuality]
      summary: List recent water quality readings
      operationId: listWaterQualityReadings
      parameters:
        - $ref: '#/components/parameters/PondIdQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/XTenantIdHeader'
      responses:
        '200': { $ref: '#/components/responses/WaterQualityReadingListResponse' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
    post:
      tags: [WaterQuality]
      summary: Create a water quality reading (auto-analyzes alerts)
      operationId: createWaterQualityReading
      parameters:
        - $ref: '#/components/parameters/XTenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWaterQualityReadingRequest'
      responses:
        '201':
          description: Reading created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaterQualityReading'
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
  /water-quality/{id}:
    get:
      tags: [WaterQuality]
      summary: Get a single water quality reading
      operationId: getWaterQualityReading
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/XTenantIdHeader'
      responses:
        '200':
          description: Reading details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaterQualityReading'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /water-quality/trends/{pondId}:
    get:
      tags: [WaterQuality]
      summary: Analyze trends for a pond over a time window
      operationId: getWaterQualityTrends
      parameters:
        - name: pondId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/DaysQuery'
        - $ref: '#/components/parameters/XTenantIdHeader'
      responses:
        '200': { description: Trend analysis returned }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /water-quality/test/mock:
    get:
      tags: [WaterQuality]
      summary: Return mock readings sample (development/testing only)
      operationId: getMockWaterQualityReadings
      responses:
        '200': { description: Mock readings returned }
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...